More pathing changes.
The script now checks for "Streamlabs Chatbot.exe" in the script_run_path. If this exists, it means Streamlabs Chatbot is running the script. If not, it means the script is being run manually. It has to use different pathing in each case. The script should now work when run manually.