Update FlagTracker_StreamlabsSystem.py
Cleaned up logging.

When channel owners removed redemptions from their Twitch dashboard, this was being detected as another redemption by the script. To avoid this, the event.Status is now being checked. If the status is Unfulfilled, it is a new redemption. If the status is Action Taken, it is the channel owner removing the redemption from their dashboard, so we don't want to add the information to the queue.