Modified Token Refresh
The creds.refresh() function was silently returning a bad request error. Apparently this is a thing that can happen at random. For now, when the program detects the token is expired, it deletes the token and opens a browser window to have the user obtain another token. Will look into a better option.